+++ test/test_random_generator/test_random_generator.cpp
@@ -15,10 +15,49 @@
  *      None.
  */
 
+#include <cstddef>
+#include <utility>
 #include <terra/random/random_generator.h>
 #include <terra/stf/stf.h>
 
 using namespace Terra::Random;
+
+// Chi-squared test to verify the histograms in the following tests
+// prove the distribution appears to be uniform
+std::pair<bool, bool> PerformChiSquaredTest(
+                                    const std::vector<std::size_t> &histogram)
+{
+    const double Threshold = 293.25;            // alpha = 0.05
+    const double Critical = 364.04;             // alpha = 0.00001
+    const double Expected = 100.0;              // Buckets have ~100 in each
+    double chi_squared = 0.0;
+    for (std::size_t count : histogram)
+    {
+        chi_squared += (count - Expected) * (count - Expected) / Expected;
+    }
+
+    return {chi_squared > Threshold, chi_squared > Critical};
+}
+
+// Check that there are not many buckets outside the expected range(this
+// has been replaced with the Chi-squared test)
+bool PerformRangeTest(const std::vector<std::size_t> &histogram)
+{
+    const std::size_t lower_bound = 65;
+    const std::size_t upper_bound = 135;
+    const std::size_t threshold = 2;
+    std::size_t count = 0;
+
+    for (auto value : histogram)
+    {
+        // Though rare, a bucket might have fewer elements -- retry
+        if (value < lower_bound) count++;
+        if (value > upper_bound) count++;
+        if (count > threshold) return false;
+    }
+
+    return true;
+}
 
 // Test to ensure that the the PRNG routine(s) are seeding distinctly
 STF_TEST(RandomGenerator, DistinctSeeding)
@@ -78,77 +117,72 @@ STF_TEST(RandomGenerator, DistinctSeedingPseudoRandom)
 // Verify the the PRNG produces a uniform distribution
 STF_TEST(RandomGenerator, UniformDistribution)
 {
-    constexpr unsigned Retry_Count = 5;
-    RandomGenerator generator;
+    constexpr std::size_t Trials = 5;
+    constexpr std::size_t Max_Failures = 4;
+    constexpr std::size_t Iterations = 25'600;
     std::vector<std::size_t> histogram(256);
-    unsigned trials;
+    std::size_t failures = 0;
 
-    // Test will be tried Retry_Count times
-    for (trials = 0; trials < Retry_Count; trials++)
+    // Test will be tried Trials times
+    for (std::size_t trials = 0; trials < Trials; trials++)
     {
-        bool retry = false;
+        RandomGenerator generator;
+
+        // Initialize the histogram
+        std::fill(histogram.begin(), histogram.end(), 0);
 
         // Generate 25'600 random octets
-        for (auto i = 0; i < 25'600; i++)
+        for (std::size_t i = 0; i < Iterations; i++)
         {
             std::uint8_t value = generator.GetRandomOctet();
             histogram[value]++;
         }
 
-        // Given a uniform distribution, each bucket of the histogram should
-        // have about 100 elements in it; assume there are at least 70
-        for (auto i = 0; i < 256; i++)
-        {
-            // Though rare, a bucket might have fewer elements. In that case,
-            // try it a second time
-            if (histogram[i] < 70) retry = true;
-            histogram[i] = 0;
-        }
-
-        // Stop if another trail is not needed
-        if (!retry) break;
+        // The histogram should have about 100 in each bucket; the following
+        // test will confirm that there is a uniform distribution
+        auto [failure, critical] = PerformChiSquaredTest(histogram);
+        if (failure) failures++;
+        STF_ASSERT_FALSE(critical);
     }
 
-    // Ensure that the trail count was not exhausted.
-    STF_ASSERT_NE(trials, Retry_Count);
+    // Ensure the number of failures is not exceeded allowable
+    STF_ASSERT_LE(failures, Max_Failures);
 }
 
 // Verify the the PRNG produces a uniform distribution (C++ PRNG only)
 STF_TEST(RandomGenerator, UniformDistributionPseudoRandom)
 {
-    constexpr unsigned Retry_Count = 5;
+    constexpr std::size_t Trials = 5;
+    constexpr std::size_t Max_Failures = 4;
+    constexpr std::size_t Iterations = 25'600;
     RandomGenerator generator(true);
     std::vector<std::size_t> histogram(256);
-    unsigned trials;
+    std::size_t failures = 0;
 
-    // Test will be tried Retry_Count times
-    for (trials = 0; trials < Retry_Count; trials++)
+    // Test will be tried Trials times
+    for (std::size_t trials = 0; trials < Trials; trials++)
     {
-        bool retry = false;
+        RandomGenerator generator;
+
+        // Initialize the histogram
+        std::fill(histogram.begin(), histogram.end(), 0);
 
         // Generate 25'600 random octets
-        for (auto i = 0; i < 25'600; i++)
+        for (std::size_t i = 0; i < Iterations; i++)
         {
             std::uint8_t value = generator.GetRandomOctet();
             histogram[value]++;
         }
 
-        // Given a uniform distribution, each bucket of the histogram should
-        // have about 100 elements in it; assume there are at least 70
-        for (auto i = 0; i < 256; i++)
-        {
-            // Though rare, a bucket might have fewer elements. In that case,
-            // try it a second time
-            if (histogram[i] < 70) retry = true;
-            histogram[i] = 0;
-        }
-
-        // Stop if another trail is not needed
-        if (!retry) break;
+        // The histogram should have about 100 in each bucket; the following
+        // test will confirm that there is a uniform distribution
+        auto [failure, critical] = PerformChiSquaredTest(histogram);
+        if (failure) failures++;
+        STF_ASSERT_FALSE(critical);
     }
 
-    // Ensure that the trail count was not exhausted.
-    STF_ASSERT_NE(trials, Retry_Count);
+    // Ensure the number of failures is not exceeded allowable
+    STF_ASSERT_LE(failures, Max_Failures);
 }
 
 // Verify the ability to retrieve a vector of random values in bulk
@@ -156,105 +190,32 @@ STF_TEST(RandomGenerator, UniformDistributionPseudoRandom)
 // overloaded functions named GetRandomOctets())
 STF_TEST(RandomGenerator, GetVectorOfRandomOctets1)
 {
-    constexpr unsigned Retry_Count = 5;
-    RandomGenerator generator(true);
+    constexpr std::size_t Trials = 5;
+    constexpr std::size_t Max_Failures = 4;
+    constexpr std::size_t Count = 25'600;
     std::vector<std::size_t> histogram(256);
-    unsigned trials;
+    std::size_t failures = 0;
 
-    // Test will be tried Retry_Count times
-    for (trials = 0; trials < Retry_Count; trials++)
+    // Test will be tried Trials times
+    for (std::size_t trials = 0; trials < Trials; trials++)
     {
-        bool retry = false;
+        RandomGenerator generator(true);
 
-        // Generate 25'600 random octets
-        std::vector<std::uint8_t> values = generator.GetRandomOctets(25'600);
-        STF_ASSERT_EQ(25'600, values.size());
+        // Initialize the histogram
+        std::fill(histogram.begin(), histogram.end(), 0);
+
+        // Generate "Count" random octets
+        std::vector<std::uint8_t> values = generator.GetRandomOctets(Count);
+        STF_ASSERT_EQ(Count, values.size());
         for (auto value : values) histogram[value]++;
 
-        // Given a uniform distribution, each bucket of the histogram should
-        // have about 100 elements in it; assume there are at least 70
-        for (auto i = 0; i < 256; i++)
-        {
-            // Though rare, a bucket might have fewer elements -- retry
-            if (histogram[i] < 70) retry = true;
-            histogram[i] = 0;
-        }
-
-        // Stop if another trail is not needed
-        if (!retry) break;
+        // The histogram should have about 100 in each bucket; the following
+        // test will confirm that there is a uniform distribution
+        auto [failure, critical] = PerformChiSquaredTest(histogram);
+        if (failure) failures++;
+        STF_ASSERT_FALSE(critical);
     }
 
-    // Ensure that the trail count was not exhausted.
-    STF_ASSERT_NE(trials, Retry_Count);
-}
-
-// Verify the ability to retrieve a vector of random values in bulk
-// and that those appear to have a uniform distribution (second of two
-// overloaded functions named GetRandomOctets())
-STF_TEST(RandomGenerator, GetVectorOfRandomOctets2)
-{
-    constexpr unsigned Retry_Count = 5;
-    RandomGenerator generator(true);
-    std::vector<std::size_t> histogram(256);
-    unsigned trials;
-
-    // Test will be tried Retry_Count times
-    for (trials = 0; trials < Retry_Count; trials++)
-    {
-        bool retry = false;
-
-        // Generate 25'600 random octets
-        auto values = generator.GetRandomOctets(25'600);
-        for (auto value : values) histogram[value]++;
-
-        // Given a uniform distribution, each bucket of the histogram should
-        // have about 100 elements in it; assume there are at least 70
-        for (auto i = 0; i < 256; i++)
-        {
-            // Though rare, a bucket might have fewer elements -- retry
-            if (histogram[i] < 70) retry = true;
-            histogram[i] = 0;
-        }
-
-        // Stop if another trail is not needed
-        if (!retry) break;
-    }
-
-    // Ensure that the trail count was not exhausted.
-    STF_ASSERT_NE(trials, Retry_Count);
-}
-
-// Verify we can fill a span of octets with random data
-STF_TEST(RandomGenerator, GetVectorOfRandomOctets3)
-{
-    constexpr unsigned Retry_Count = 5;
-    RandomGenerator generator(true);
-    std::vector<std::size_t> histogram(256);
-    unsigned trials;
-
-    // Test will be tried Retry_Count times
-    for (trials = 0; trials < Retry_Count; trials++)
-    {
-        bool retry = false;
-
-        // Generate 25'600 random octets
-        std::vector<std::uint8_t> values(25'600);
-        generator.GetRandomOctets(values);
-        for (auto value : values) histogram[value]++;
-
-        // Given a uniform distribution, each bucket of the histogram should
-        // have about 100 elements in it; assume there are at least 70
-        for (auto i = 0; i < 256; i++)
-        {
-            // Though rare, a bucket might have fewer elements -- retry
-            if (histogram[i] < 70) retry = true;
-            histogram[i] = 0;
-        }
-
-        // Stop if another trail is not needed
-        if (!retry) break;
-    }
-
-    // Ensure that the trail count was not exhausted.
-    STF_ASSERT_NE(trials, Retry_Count);
+    // Ensure the number of failures is not exceeded allowable
+    STF_ASSERT_LE(failures, Max_Failures);
 }